requirements.txt
openai>=1.0.0
telethon>=1.34.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
google-api-python-client
langchain
langchain-openai
langchain-core

# Bot dependencies
aiogram==3.5.0
apscheduler==3.10.4
SQLAlchemy==2.0.29
asyncpg==0.29.0
psycopg2-binary==2.9.9
